"Some", not "All", contributors
One problem with "All contributors" is that it's not all contributors and it de-values the contributions of contributors further from the code. UX professionals including Maria Dahman have contributed, but are not credited.  Product and service management roles including Jim Helwig and Annette Stratman-Durrer have contributed, but are not credited. 

This change doesn't fully fix this, but at least it starts to acknowledge that this is partial rather than  comprehensive list of contributors.